1)Rajout des fonctions manquantes dans les .h, par rapport à l'UML. 2)Modification du Constructeur Puzzle. 3)Modification de la fonction Theory dans la classe Euclidean, afin qu'elle compile sur windows. 4)Petite modif de l'UML. 5)Concombre

+++ trunk/Taquin42/Taquin42/Euclidean.hpp
@@ -19,11 +19,11 @@ private:
   Euclidean &				operator=(const Euclidean &);
 
 public:
-  inline int				Theory(const sPositions & NodePos,
+  inline double				Theory(const sPositions & NodePos,
 				       	const sGoalPositions & GoalNodePos)
   {
-  	double	dx = pow((NodePos.Node_px - GoalNodePos.GoalNode_px), 2);
-  	double	dy = pow((NodePos.Node_py - GoalNodePos.GoalNode_py), 2);
+ 	double	dx = pow(static_cast<double>(NodePos.Node_px - GoalNodePos.GoalNode_px), 2);
+ 	double	dy = pow(static_cast<double>(NodePos.Node_py - GoalNodePos.GoalNode_py), 2);
   	return (sqrt(dx + dy));
   }
